# Speed up bond repository update and remove some raw types