--- a/core/src/main/java/bisq/core/dao/governance/bond/BondRepository.java
+++ b/core/src/main/java/bisq/core/dao/governance/bond/BondRepository.java
@@ -56,7 +56,7 @@ import lombok.extern.slf4j.Slf4j;
  * unconfirmed txs.
  */
 @Slf4j
-public abstract class BondRepository<T extends Bond, R extends BondedAsset> implements DaoSetupService,
+public abstract class BondRepository<B extends Bond<T>, T extends BondedAsset> implements DaoSetupService,
         BsqWalletService.WalletTransactionsChangeListener {
 
     ///////////////////////////////////////////////////////////////////////////////////////////
@@ -64,7 +64,7 @@ public abstract class BondRepository<T extends Bond, R extends BondedAsset> impl
     ///////////////////////////////////////////////////////////////////////////////////////////
 
     public static void applyBondState(DaoStateService daoStateService,
-                                      Bond bond,
+                                      Bond<?> bond,
                                       Tx lockupTx,
                                       TxOutput lockupTxOutput) {
         if (bond.getBondState() != BondState.LOCKUP_TX_PENDING || bond.getBondState() != BondState.UNLOCK_TX_PENDING)
@@ -132,7 +132,7 @@ public abstract class BondRepository<T extends Bond, R extends BondedAsset> impl
                 .anyMatch(data -> Arrays.equals(BondConsensus.getHashFromOpReturnData(data), bondedAsset.getHash()));
     }
 
-    public static boolean isConfiscated(Bond bond, DaoStateService daoStateService) {
+    public static boolean isConfiscated(Bond<?> bond, DaoStateService daoStateService) {
         return (bond.getLockupTxId() != null && daoStateService.isConfiscatedLockupTxOutput(bond.getLockupTxId())) ||
                 (bond.getUnlockTxId() != null && daoStateService.isConfiscatedUnlockTxOutput(bond.getUnlockTxId()));
     }
@@ -142,9 +142,9 @@ public abstract class BondRepository<T extends Bond, R extends BondedAsset> impl
     protected final BsqWalletService bsqWalletService;
 
     // This map is just for convenience. The data which are used to fill the map are stored in the DaoState (role, txs).
-    protected final Map<String, T> bondByUidMap = new HashMap<>();
+    protected final Map<String, B> bondByUidMap = new HashMap<>();
     @Getter
-    protected final ObservableList<T> bonds = FXCollections.observableArrayList();
+    protected final ObservableList<B> bonds = FXCollections.observableArrayList();
 
 
     ///////////////////////////////////////////////////////////////////////////////////////////
@@ -192,12 +192,12 @@ public abstract class BondRepository<T extends Bond, R extends BondedAsset> impl
     // API
     ///////////////////////////////////////////////////////////////////////////////////////////
 
-    public boolean isBondedAssetAlreadyInBond(R bondedAsset) {
+    public boolean isBondedAssetAlreadyInBond(T bondedAsset) {
         boolean contains = bondByUidMap.containsKey(bondedAsset.getUid());
         return contains && bondByUidMap.get(bondedAsset.getUid()).getLockupTxId() != null;
     }
 
-    public List<Bond> getActiveBonds() {
+    public List<B> getActiveBonds() {
         return bonds.stream().filter(Bond::isActive).collect(Collectors.toList());
     }
 
@@ -206,17 +206,17 @@ public abstract class BondRepository<T extends Bond, R extends BondedAsset> impl
     // Protected
     ///////////////////////////////////////////////////////////////////////////////////////////
 
-    protected abstract T createBond(R bondedAsset);
+    protected abstract B createBond(T bondedAsset);
 
-    protected abstract void updateBond(T bond, R bondedAsset, TxOutput lockupTxOutput);
+    protected abstract void updateBond(B bond, T bondedAsset, TxOutput lockupTxOutput);
 
-    protected abstract Stream<R> getBondedAssetStream();
+    protected abstract Stream<T> getBondedAssetStream();
 
     protected void update() {
         log.debug("update");
         getBondedAssetStream().forEach(bondedAsset -> {
             String uid = bondedAsset.getUid();
-            T bond = bondByUidMap.computeIfAbsent(uid, u -> createBond(bondedAsset));
+            B bond = bondByUidMap.computeIfAbsent(uid, u -> createBond(bondedAsset));
 
             daoStateService.getLockupTxOutputs().forEach(lockupTxOutput ->
                     updateBond(bond, bondedAsset, lockupTxOutput));