--- a/core/src/test/java/bisq/core/dao/governance/proposal/MyProposalListServiceTest.java
+++ b/core/src/test/java/bisq/core/dao/governance/proposal/MyProposalListServiceTest.java
@@ -11,20 +11,24 @@ import bisq.common.persistence.PersistenceManager;
 
 import javafx.beans.property.SimpleIntegerProperty;
 
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+@ExtendWith(MockitoExtension.class)
 public class MyProposalListServiceTest {
     @Test
-    public void canInstantiate() {
+    public void canInstantiate(@Mock PersistenceManager<MyProposalList> persistenceManager) {
         P2PService p2PService = mock(P2PService.class);
         when(p2PService.getNumConnectedPeers()).thenReturn(new SimpleIntegerProperty(0));
-        PersistenceManager persistenceManager = mock(PersistenceManager.class);
-        MyProposalListService service = new MyProposalListService(p2PService,
-                mock(DaoStateService.class),
-                mock(PeriodService.class), mock(WalletsManager.class), persistenceManager, mock(PubKeyRing.class)
+
+        new MyProposalListService(p2PService, mock(DaoStateService.class), mock(PeriodService.class),
+                mock(WalletsManager.class), persistenceManager, mock(PubKeyRing.class)
         );
     }
 }